# Speed up compilation of large constant arrays