@nnethercote 2018-06-27
Nice work!

> Edit: Inlining Size::bytes() saves an additional 6 seconds dropping the total time to 13 seconds on my machine.

For the first patch in #51672 I inlined several methods in addition to `bytes()`. This was based on profiling data -- while `bytes()` is clearly the hottest, the others were still hot, and I think it would be worth trying to inline all of them to see if you get a bigger speed-up.