@michaelwoerister 2018-06-27
It would be interesting to see if this could be further sped up by copying larger chunks of memory at a time. Right now this makes `n` calls to `memcpy` but `memcpy` can be more efficient when it works with more memory at once (via SIMD). We could try increase the chunk size and see what that does to performance. In particular copy `[0 .. size] to [size .. size *2]`, then `[0 .. size*2]` to `[size*2 .. size*4]`, then `[0..size*4]` to `[size*4 .. size*8]`, etc, up to a maximum chunk size. This makes `copy_repeatedly` quite a bit more complicated but it might be worth it?