@scottmcm 2018-06-28
> increase the chunk size

Looks like the logic for that already exists in core:

https://github.com/rust-lang/rust/blob/3515dab431702f3a926f61f52cd0743d95e4fd3f/src/liballoc/slice.rs#L422

Maybe there's a way to extract the copying part into an unsafe `ptr` method and re-use it here?