@wesleywiser 2018-06-28
@nnethercote That's a good idea. I tried that and it saved an additional 3 seconds on the compile time (now 10 seconds).