@wesleywiser 2018-06-30
@michaelwoerister @scottmcm I tried that but it only saved about 100ms. `perf` indicates that the memory copy operations aren't hot. Good idea though!