@wesleywiser 2018-06-30
With the latest changes, the compile time is down to 2.8 seconds on my machine.