@oli-obk 2018-06-30
Please also add a test for http://play.rust-lang.org/?gist=1c0e90ac9064edfa12fbd286902e20ef to make sure we always properly copy the relocations.