@wesleywiser 2018-07-01
Fixed tidy