@oli-obk 2018-07-01
@bors r+

We should probably add a bunch of tests to the perf tests to ensure this doesn't regress.