@michaelwoerister 2018-07-02
> @michaelwoerister @scottmcm I tried that but it only saved about 100ms. perf indicates that the memory copy operations aren't hot. Good idea though!

Thanks for giving it a try, @wesleywiser. Great work!